image-search/Lambda/classifydata/requirements.txt
boto3~=1.34.84
pymongo~=4.6.3
python-dotenv~=1.0.1
certifi~=2024.2.2

botocore~=1.34.84
